moved to tests/lib
