--- tests/constants.py
@@ -1,32 +1,0 @@
-#!/usr/bin/python3
-# -*- coding: utf-8 -*-
-# input-remapper - GUI for device specific keyboard mappings
-# Copyright (C) 2022 sezanzeb <proxima@sezanzeb.de>
-#
-# This file is part of input-remapper.
-#
-# input-remapper is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# input-remapper is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with input-remapper.  If not, see <https://www.gnu.org/licenses/>.
-
-# give tests some time to test stuff while the process
-# is still running
-EVENT_READ_TIMEOUT = 0.01
-
-# based on experience how much time passes at most until
-# the reader-service starts receiving previously pushed events after a
-# call to start_reading
-START_READING_DELAY = 0.05
-
-# for joysticks
-MIN_ABS = -(2**15)
-MAX_ABS = 2**15
