+++ tests/integration/test_numlock.py
@@ -19,7 +19,7 @@
 # along with input-remapper.  If not, see <https://www.gnu.org/licenses/>.
 
 
-from tests.cleanup import quick_cleanup
+from tests.lib.cleanup import quick_cleanup
 
 import unittest
