--- tests/stuff.py
@@ -1,38 +1,0 @@
-#!/usr/bin/python3
-# -*- coding: utf-8 -*-
-# input-remapper - GUI for device specific keyboard mappings
-# Copyright (C) 2022 sezanzeb <proxima@sezanzeb.de>
-#
-# This file is part of input-remapper.
-#
-# input-remapper is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# input-remapper is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with input-remapper.  If not, see <https://www.gnu.org/licenses/>.
-
-import os
-import copy
-from unittest.mock import patch
-
-
-def convert_to_internal_events(events):
-    """Convert an iterable of InputEvent to a list of inputremapper.InputEvent."""
-    from inputremapper.input_event import InputEvent as InternalInputEvent
-
-    return [InternalInputEvent.from_event(event) for event in events]
-
-
-def spy(obj, name):
-    """Convenient wrapper for patch.object(..., ..., wraps=...)."""
-    return patch.object(obj, name, wraps=obj.__getattribute__(name))
-
-
-environ_copy = copy.deepcopy(os.environ)
