+++ tests/unit/test_context.py
@@ -18,8 +18,8 @@
 # You should have received a copy of the GNU General Public License
 # along with input-remapper.  If not, see <https://www.gnu.org/licenses/>.
 
-from tests.cleanup import quick_cleanup
-from tests.fixtures import get_key_mapping
+from tests.lib.cleanup import quick_cleanup
+from tests.lib.fixtures import get_key_mapping
 from evdev.ecodes import (
     EV_REL,
     EV_ABS,
