+++ tests/unit/test_control.py
@@ -22,8 +22,8 @@
 """Testing the input-remapper-control command"""
 
 
-from tests.cleanup import quick_cleanup
-from tests.tmp import tmp
+from tests.lib.cleanup import quick_cleanup
+from tests.lib.tmp import tmp
 
 import os
 import time
