+++ tests/unit/test_event_pipeline/test_event_pipeline.py
@@ -59,11 +59,11 @@ from inputremapper.injection.context import Context
 from inputremapper.injection.event_reader import EventReader
 from inputremapper.injection.global_uinputs import global_uinputs
 from inputremapper.input_event import InputEvent, USE_AS_ANALOG_VALUE
-from tests.cleanup import cleanup
-from tests.logger import logger
-from tests.constants import MAX_ABS, MIN_ABS
-from tests.stuff import convert_to_internal_events
-from tests.fixtures import (
+from tests.lib.cleanup import cleanup
+from tests.lib.logger import logger
+from tests.lib.constants import MAX_ABS, MIN_ABS
+from tests.lib.stuff import convert_to_internal_events
+from tests.lib.fixtures import (
     Fixture,
     fixtures,
     get_key_mapping,
