+++ tests/unit/test_event_pipeline/test_mapping_handlers.py
@@ -56,10 +56,10 @@ from inputremapper.injection.mapping_handlers.macro_handler import MacroHandler
 from inputremapper.injection.mapping_handlers.mapping_handler import MappingHandler
 from inputremapper.injection.mapping_handlers.rel_to_abs_handler import RelToAbsHandler
 from inputremapper.input_event import InputEvent, EventActions, USE_AS_ANALOG_VALUE
-from tests.cleanup import cleanup
-from tests.patches import InputDevice
-from tests.constants import MAX_ABS
-from tests.stuff import convert_to_internal_events
+from tests.lib.cleanup import cleanup
+from tests.lib.patches import InputDevice
+from tests.lib.constants import MAX_ABS
+from tests.lib.stuff import convert_to_internal_events
 
 
 class BaseTests:
