+++ tests/unit/test_event_reader.py
@@ -43,9 +43,9 @@ from inputremapper.event_combination import EventCombination
 from inputremapper.injection.context import Context
 from inputremapper.injection.event_reader import EventReader
 from inputremapper.injection.global_uinputs import global_uinputs
-from tests.fixtures import new_event
-from tests.cleanup import quick_cleanup
-from tests.fixtures import get_key_mapping
+from tests.lib.fixtures import new_event
+from tests.lib.cleanup import quick_cleanup
+from tests.lib.fixtures import get_key_mapping
 
 
 class TestEventReader(unittest.IsolatedAsyncioTestCase):
