+++ tests/unit/test_global_uinputs.py
@@ -19,7 +19,7 @@
 # along with input-remapper.  If not, see <https://www.gnu.org/licenses/>.
 
 
-from tests.cleanup import cleanup
+from tests.lib.cleanup import cleanup
 
 import sys
 import unittest
