+++ tests/unit/test_ipc.py
@@ -20,8 +20,8 @@
 import asyncio
 import multiprocessing
 
-from tests.cleanup import quick_cleanup
-from tests.tmp import tmp
+from tests.lib.cleanup import quick_cleanup
+from tests.lib.tmp import tmp
 
 import unittest
 import select
