+++ tests/unit/test_logger.py
@@ -19,7 +19,7 @@
 # along with input-remapper.  If not, see <https://www.gnu.org/licenses/>.
 
 
-from tests.tmp import tmp
+from tests.lib.tmp import tmp
 
 import os
 import shutil
